add ItemsPicker component
