+++ ts/features/bonus/siciliaVola/screens/voucherGeneration/StudentSelectDestinationScreen.tsx
@@ -1,8 +1,7 @@
 import * as React from "react";
-import { useRef, useState } from "react";
 import { connect } from "react-redux";
 import { Dispatch } from "redux";
-import { SafeAreaView, ScrollView, StyleSheet } from "react-native";
+import { SafeAreaView, ScrollView } from "react-native";
 import { isSome } from "fp-ts/lib/Option";
 import BaseScreenComponent from "../../../../../components/screens/BaseScreenComponent";
 import { emptyContextualHelp } from "../../../../../utils/emptyContextualHelp";
@@ -20,32 +19,14 @@ import { University } from "../../types/SvVoucherRequest";
 import { selectedBeneficiaryCategorySelector } from "../../store/reducers/voucherRequest";
 import { navigateToSvSelectFlightsDateScreen } from "../../navigation/actions";
 import I18n from "../../../../../i18n";
-import { Picker } from "native-base";
-import IconFont from "../../../../../components/ui/IconFont";
-import variables from "../../../../../theme/variables";
-import { IOColors } from "../../../../../components/core/variables/IOColors";
+import ItemsPicker from "../../../../../components/ui/ItemsPicker";
 
 type Props = ReturnType<typeof mapDispatchToProps> &
   ReturnType<typeof mapStateToProps>;
 
-const styles = StyleSheet.create({
-  container: { borderBottomWidth: 1, borderColor: IOColors.bluegreyLight },
-  inputContainer: {
-    flex: 1,
-    flexDirection: "row",
-    justifyContent: "space-between"
-  },
-  icon: {
-    height: 24,
-    width: 24,
-    marginBottom: 5
-  }
-});
-
 const StudentSelectDestinationScreen = (
   props: Props
 ): React.ReactElement | null => {
-  const elementRef = useRef(null);
   const backButtonProps = {
     primary: false,
     bordered: true,
@@ -66,7 +47,6 @@ const StudentSelectDestinationScreen = (
     props.failure("The selected category is not Student");
     return null;
   }
-  const [selectedValue, setSelectedValue] = useState();
   return (
     <BaseScreenComponent
       goBack={true}
@@ -76,7 +56,6 @@ const StudentSelectDestinationScreen = (
       <SafeAreaView
         style={IOStyles.flex}
         testID={"StudentSelectDestinationScreen"}
-        ref={elementRef}
       >
         <ScrollView style={[IOStyles.horizontalContentPadding]}>
           <H1>
@@ -84,48 +63,11 @@ const StudentSelectDestinationScreen = (
               "bonus.sv.voucherGeneration.student.selectDestination.title"
             )}
           </H1>
-          <Picker
-            /*style={{ position: "absolute", bottom: 0, left: 0, right: 0 }}*/
-            enabled={true}
-            mode={"dropdown"}
-            collapsable={true}
-            selectedValue={selectedValue}
-            onValueChange={(itemValue, _) => setSelectedValue(itemValue)}
-            headerTitleStyle={{
-              fontFamily: "Titillium Web",
-              color: "black",
-              textAlign: "auto",
-              paddingLeft: 15
-            }}
-            headerBackButtonTextStyle={{
-              fontFamily: "Titillium Web",
-              color: "black",
-              fontSize: 16,
-              textAlign: "auto",
-              lineHeight: undefined,
-              paddingLeft: 0,
-              fontWeight: "300"
-            }}
-            placeholder={"seleziona una scelta"}
-            iosIcon={
-              <IconFont
-                size={variables.iconSize3}
-                color={IOColors.blue}
-                name={"io-calendario"}
-                selectionColor={IOColors.blue}
-                style={styles.icon}
-              />
-            }
-            style={{
-              flex: 1,
-              justifyContent: "space-between",
-              alignSelf: "auto"
-            }}
-          >
-            <Picker.Item label="Office eg: VC, DSA" value="" color="#c1c1c1" />
-            <Picker.Item label={"pippo"} value={"pippo"} />
-            <Picker.Item label={"pluto"} value={"pluto"} />
-          </Picker>
+          <ItemsPicker
+            placeholder={"Seleziona un comune"}
+            items={[]}
+            // onValueChange={(a, b) => true}
+          />
         </ScrollView>
 
         <FooterWithButtons
